requirements.txt
streamlit==1.35.0
numpy==1.26.4
opencv-python-headless==4.9.0.80
Pillow==10.3.0
matplotlib==3.8.4
seaborn==0.13.2
scikit-learn==1.4.2
PyWavelets==1.6.0
xgboost==2.0.3
tensorflow==2.16.1
efficientnet==1.1.1
